rearrange redux index - website
